changing epic5 to epic5twitch in useful info broke scripts that checked to make sure the client is epic5. So changed it back.
Also changed version to 2.1.0 instead of 0.1.0 because it's better looking, plus it is more representative that it actually is epic5 2.0.1 and higher.. There is no epic5 2.1.0 as far as I know, and even if there is, this is epic5twitch BUT it still identifies as epic5 because it's just a fork of epic5, not a whole new client.

Basically epic5 with Twitch support.